requirements.txt
PyGObject>=3.42.0
reportlab>=4.0.0
Pillow>=10.0.0
